api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.3
psycopg2-binary==2.9.9

python-dotenv==1.0.1